use arrays instead of objects - constructor names are not relevant during the evaluation
generated JS code before:

    exports["Bool"]["false"] = x0 => x0["false"]();
    exports["Bool"]["true"] = x0 => x0["true"]();
    exports["not"] = x0 => x0({
        "false": () => exports["Bool"]["true"],
        "true": () => exports["Bool"]["false"]
      });

generated JS code after:

    exports["Bool"]["false"] = x0 => x0[0]();
    exports["Bool"]["true"] = x0 => x0[1]();
    exports["not"] = x0 => x0([
        () => exports["Bool"]["true"],
        () => exports["Bool"]["false"]
      ]);

Erased branches are replaced by `null`.
If more than the half of branches are erased, the array is compressed to
be a object like {3: ..., 13: ...}.

The patch contains a little refactoring step: the MemberIndex data type
is merged into the MemberId data type.