dead code elimination for the JS backend
Dead code elimination works only for concatenated JS output, i.e. if
`agda` is called with `--js-output`.

Dead code elimination will keep the definitions which are used
(transitively) in the definition of external symbols.

The new flag `--js-externals` tells the JS backend where to find the
file which contains the list of external symbols.

An example text file listing external symbols:

   Prelude.natToFin
   Prelude.Bool.true
   Prelude.Bool.false
   Prelude.Maybe.just
   Reactive.WidgetEdit.modLeft

If the list is empty or no `--js-external` are given then all definitons
are kept.