rename top level definitions
Renaminging is done only if external identifiers are provided with
`--js-externals`.

Top level definitions will be renamed to two-character names plus an
optional number:

    aa, ab, ac, ..., ba, bb, bc, ..., aa1, ab1, ac1, ...

External identifiers are renamed to the given name in the file which
lists external identifiers.

The new format of the external identifiers file is:
Each row contains an external identifier and its new name separated by a
space character.
For example:

   Prelude.natToFin natToFin
   Prelude.Bool.true agda_true
   Reactive.WidgetEdit.modLeft modLeft